% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{add_helper_columns}
\alias{add_helper_columns}
\title{Add helper columns to assessment data}
\usage{
add_helper_columns(df)
}
\arguments{
\item{df}{Data frame with assessment data}
}
\value{
Data frame with helper columns added
}
\description{
Adds is_state, is_district, is_school helper columns based on
school_code and district_code presence.
}
\keyword{internal}
