% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{construct_lss_enrollment_from_published}
\alias{construct_lss_enrollment_from_published}
\title{Construct LSS enrollment from published totals}
\usage{
construct_lss_enrollment_from_published(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with LSS enrollment
}
\description{
Creates district-level enrollment data from MSDE published totals.
This is a fallback when detailed data is not available.
}
\keyword{internal}
