% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{create_empty_processed_df}
\alias{create_empty_processed_df}
\title{Create empty processed data frame}
\usage{
create_empty_processed_df(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Empty data frame with expected structure
}
\description{
Create empty processed data frame
}
\keyword{internal}
