% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_mcap_file}
\alias{download_mcap_file}
\title{Download MCAP data file from URL}
\usage{
download_mcap_file(url, end_year)
}
\arguments{
\item{url}{URL to download}

\item{end_year}{School year end}
}
\value{
Data frame with assessment data
}
\description{
Downloads and extracts MCAP data from a ZIP file on the Report Card.
}
\keyword{internal}
