% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_md_reportcard_enrollment}
\alias{download_md_reportcard_enrollment}
\title{Download Maryland Report Card enrollment data}
\usage{
download_md_reportcard_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with enrollment data
}
\description{
Attempts to fetch enrollment data from the Maryland Report Card website.
This function handles the web interaction required to download data.
}
\examples{
\dontrun{
enr <- download_md_reportcard_enrollment(2024)
}
}
