% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_mdp_enrollment_file}
\alias{download_mdp_enrollment_file}
\title{Download MD Planning enrollment Excel file}
\usage{
download_mdp_enrollment_file(release_year)
}
\arguments{
\item{release_year}{The MDP release year}
}
\value{
Path to downloaded file, or NULL if download failed
}
\description{
Download MD Planning enrollment Excel file
}
\keyword{internal}
