% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_msde_enrollment_publication}
\alias{download_msde_enrollment_publication}
\title{Download enrollment data from MSDE Staff and Student Publications}
\usage{
download_msde_enrollment_publication(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with enrollment data
}
\description{
Downloads enrollment PDF publications from MSDE and extracts data.
Publications are available at:
https://marylandpublicschools.org/about/Documents/DCAA/SSP/
}
\keyword{internal}
