% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_school_enrollment_file}
\alias{download_school_enrollment_file}
\title{Download school enrollment from Report Card data files}
\usage{
download_school_enrollment_file(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with school enrollment
}
\description{
Download school enrollment from Report Card data files
}
\keyword{internal}
