% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{ensure_standard_columns}
\alias{ensure_standard_columns}
\title{Ensure all standard columns exist}
\usage{
ensure_standard_columns(df, end_year)
}
\arguments{
\item{df}{Data frame to check}

\item{end_year}{School year end}
}
\value{
Data frame with all standard columns
}
\description{
Ensure all standard columns exist
}
\keyword{internal}
