% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{extract_jurisdiction_grades}
\alias{extract_jurisdiction_grades}
\title{Extract grade-level enrollment for a jurisdiction}
\usage{
extract_jurisdiction_grades(df, j_row, year_col, end_year)
}
\arguments{
\item{df}{The Excel data frame}

\item{j_row}{Row number of the jurisdiction header}

\item{year_col}{Column number containing the target year's data}

\item{end_year}{The school year end}
}
\value{
Data frame with one row containing grade-level enrollment
}
\description{
Extract grade-level enrollment for a jurisdiction
}
\keyword{internal}
