% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{fetch_lss_enrollment}
\alias{fetch_lss_enrollment}
\title{Fetch LSS (District) level enrollment data}
\usage{
fetch_lss_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with district-level enrollment
}
\description{
Downloads enrollment data for all 24 Local School Systems in Maryland.
}
\keyword{internal}
