% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{fetch_reportcard_demographics}
\alias{fetch_reportcard_demographics}
\title{Fetch demographics from Report Card (if available)}
\usage{
fetch_reportcard_demographics(end_year, level = "lea")
}
\arguments{
\item{end_year}{School year end}

\item{level}{Level: "state", "lea", or "school"}
}
\value{
Data frame with demographics data
}
\description{
Fetch demographics from Report Card (if available)
}
\keyword{internal}
