% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{fetch_school_enrollment}
\alias{fetch_school_enrollment}
\title{Fetch school-level enrollment data}
\usage{
fetch_school_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with school-level enrollment
}
\description{
Downloads enrollment data for all schools in Maryland.
}
\keyword{internal}
