% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_grade_codes}
\alias{get_grade_codes}
\title{Get Maryland grade level codes}
\usage{
get_grade_codes()
}
\value{
Character vector of grade codes
}
\description{
Returns the standard grade level codes used in Maryland enrollment data.
}
\keyword{internal}
