% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{grade_to_column}
\alias{grade_to_column}
\title{Convert grade code to column name}
\usage{
grade_to_column(grade)
}
\arguments{
\item{grade}{Grade code from API}
}
\value{
Column name or NULL if unknown grade
}
\description{
Convert grade code to column name
}
\keyword{internal}
