% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{merge_enrollment_demographics}
\alias{merge_enrollment_demographics}
\title{Merge enrollment data with demographics}
\usage{
merge_enrollment_demographics(enr_data, demo_data)
}
\arguments{
\item{enr_data}{Data frame with enrollment by grade}

\item{demo_data}{Data frame with demographic breakdowns}
}
\value{
Merged data frame
}
\description{
Combines MD Planning enrollment (by grade) with MSDE demographic data.
}
\keyword{internal}
