% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{parse_mdp_enrollment_xlsx}
\alias{parse_mdp_enrollment_xlsx}
\title{Parse MD Planning enrollment Excel file}
\usage{
parse_mdp_enrollment_xlsx(xlsx_path, end_year)
}
\arguments{
\item{xlsx_path}{Path to the Excel file}

\item{end_year}{The school year end to extract}
}
\value{
Data frame with enrollment data
}
\description{
Parses the 3-Public-School-Enrollment.xlsx file from MD Planning.
The file has a complex structure with blocks for each jurisdiction.
}
\keyword{internal}
