% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_generic_enrollment}
\alias{process_generic_enrollment}
\title{Process generic enrollment data}
\usage{
process_generic_enrollment(df, end_year)
}
\arguments{
\item{df}{Raw data frame}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Attempts to process enrollment data with unknown structure.
}
\keyword{internal}
