% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{process_raw_assessment}
\alias{process_raw_assessment}
\title{Process raw assessment data}
\usage{
process_raw_assessment(df, end_year)
}
\arguments{
\item{df}{Raw data frame from download}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Standardizes column names and adds metadata columns.
}
\keyword{internal}
