% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_school_data}
\alias{process_school_data}
\title{Process school-level enrollment data}
\usage{
process_school_data(df, end_year)
}
\arguments{
\item{df}{Raw school data frame}

\item{end_year}{School year end}
}
\value{
Processed school data frame
}
\description{
Process school-level enrollment data
}
\keyword{internal}
