% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_typed_data}
\alias{process_typed_data}
\title{Process data that already has type column}
\usage{
process_typed_data(df, end_year)
}
\arguments{
\item{df}{Data frame with type column}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Process data that already has type column
}
\keyword{internal}
