% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{read_assessment_csv}
\alias{read_assessment_csv}
\title{Read assessment CSV file}
\usage{
read_assessment_csv(file_path, end_year)
}
\arguments{
\item{file_path}{Path to CSV file}

\item{end_year}{School year end}
}
\value{
Data frame with parsed assessment data
}
\description{
Reads and parses an assessment CSV file downloaded from Maryland Report Card.
}
\keyword{internal}
