% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_md_reportcard_enrollment}
\alias{download_md_reportcard_enrollment}
\title{Download enrollment data from Maryland Report Card}
\usage{
download_md_reportcard_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with enrollment data
}
\description{
Downloads enrollment data from the Maryland Report Card system.
The Report Card provides JSON data that we convert to a data frame.
}
\keyword{internal}
