% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{fetch_reportcard_demographics}
\alias{fetch_reportcard_demographics}
\title{Fetch demographics data from Maryland Report Card}
\usage{
fetch_reportcard_demographics(end_year, level = "district")
}
\arguments{
\item{end_year}{School year end}

\item{level}{One of "state", "district", or "school"}
}
\value{
Data frame with demographic data
}
\description{
Queries the Maryland Report Card system for demographic/enrollment data.
}
\keyword{internal}
