% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from MSDE}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024)}
}
\value{
Data frame with raw enrollment data
}
\description{
Downloads enrollment data from the Maryland Report Card system.
Uses the appropriate method based on the year requested.
}
\keyword{internal}
